backend/requirements.txt
fastapi
uvicorn
python-chess
stockfish
zstandard